Compare the gamestate pool assert operands as integers
GCC's -Wenum-compare (an error under -Werror on the Linux release builds)
rejects comparing constants that belong to two different anonymous enums.
Cast both sides of the pool-vs-message bound to int32_t; the check itself
is unchanged.

+++ src/qcommon/game_state_types.h
@@ -52,7 +52,7 @@ enum { MAX_GAMESTATE_CHARS = MAX_GAMESTATE_CHARS_EXTENDED };
  * pool above MAX_MSGLEN is safe (it is client-internal, never on the wire)
  * and only adds headroom for configstring growth during play. */
 GAME_STATE_TYPES_STATIC_ASSERT(
-    MAX_GAMESTATE_CHARS >= MAX_MSGLEN,
+    (int32_t)MAX_GAMESTATE_CHARS >= (int32_t)MAX_MSGLEN,
     "the gamestate pool must retain any single-message gamestate");
 
 /*
